Removes purge task
When running locally - this adds a ton of overhead fetching dependencies
after each run. The purge was a byproduct of an older docker image
testing suite that is no longer in use, and can therefore be safely
retired. CI will be responsible for always testing with a fresh suite of
dependencies.

see the drone.yaml for more info, and the CI result logs on Drone.io