Adds github-event-listener, and welcome scripts
The event-listener appears to be a more robust github integration,
emitting data for all the optional notification types to a specific
room.

https://www.npmjs.com/package/hubot-github-event-announcer

The systemzoo-welcome package is a direct fork of hubot-welcome with
some scoping to room that should help limit scrappies spam, and only
announce in the room that scrappie should be monitoring for new users.